pyproject.toml
[project]
name = "halo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "transformers>=4.45.0",
    "torch>=2.0.0",
    "torchvision",
    "pillow",
    "accelerate",
    "qwen-vl-utils",
    "python-dotenv",
    "huggingface-hub",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "psutil>=6.0.0",
    "requests>=2.32.0",
    "google-generativeai>=0.3.0",
    "prompt_toolkit>=3.0.0",
    "paho-mqtt>=2.0.0",
    "sentence-transformers>=2.0.0",
    "numpy>=1.24.0",
    "spacy>=3.7.0",
    "trl>=0.25.0",
    "datasets>=4.4.0",
    "google-genai>=0.2.0",
]

[project.scripts]
halo-chat = "halo.chat:main"
